:beetle: fixed #20 Project 페이지 렌더링 시, 아이콘 일부 렌더링 안되는 버그 수정 완료
