+++ components/Projects/Item.tsx
@@ -6,14 +6,6 @@ import { styled } from '@mui/material/styles';
 import ModuleCard from 'modules/Card';
 import theme from '../../config/theme';
 
-const Item = styled(Paper)(({ theme }) => ({
-  backgroundColor: 'dark',
-  // ...theme?.typography?.body2,
-  padding: theme.spacing(1),
-  textAlign: 'center',
-  minHeight: '100%',
-}));
-
 const moveToLink = (href: string) => {
   window.open(href, '_blank');
 };
