--- components/analytics.tsx
@@ -1,7 +1,0 @@
-'use client';
-
-import { Analytics } from '@vercel/analytics/react';
-
-export default function AnalyticsWrapper() {
-  return <Analytics />;
-}
