--- lib/twitter.ts
@@ -1,83 +1,0 @@
-// react-static-tweets isn't currently working
-// https://github.com/transitive-bullshit/react-static-tweets/issues/55
-// so using the API directly for now, maybe change later
-//
-// You can view previous version with react-static-tweets in the PR commits
-// https://github.com/leerob/leerob.io/pull/564
-export const getTweets = async (ids) => {
-  if (ids.length === 0) {
-    return [];
-  }
-
-  const queryParams = new URLSearchParams({
-    ids: ids.join(','),
-    expansions:
-      'author_id,attachments.media_keys,referenced_tweets.id,referenced_tweets.id.author_id',
-    'tweet.fields':
-      'attachments,author_id,public_metrics,created_at,id,in_reply_to_user_id,referenced_tweets,text',
-    'user.fields': 'id,name,profile_image_url,protected,url,username,verified',
-    'media.fields':
-      'duration_ms,height,media_key,preview_image_url,type,url,width,public_metrics',
-  });
-
-  const response = await fetch(
-    `https://api.twitter.com/2/tweets?${queryParams}`,
-    {
-      headers: {
-        Authorization: `Bearer ${process.env.TWITTER_API_TOKEN}`,
-      },
-    }
-  );
-
-  const tweets = await response.json();
-
-  const getAuthorInfo = (author_id) => {
-    return tweets.includes.users.find((user) => user.id === author_id);
-  };
-
-  const getReferencedTweets = (mainTweet) => {
-    return (
-      mainTweet?.referenced_tweets?.map((referencedTweet) => {
-        const fullReferencedTweet = tweets.includes.tweets.find(
-          (tweet) => tweet.id === referencedTweet.id
-        );
-
-        return {
-          type: referencedTweet.type,
-          author: getAuthorInfo(fullReferencedTweet.author_id),
-          ...fullReferencedTweet,
-        };
-      }) || []
-    );
-  };
-
-  return (
-    tweets.data.reduce((allTweets, tweet) => {
-      const tweetWithAuthor = {
-        ...tweet,
-        media:
-          tweet?.attachments?.media_keys.map((key) =>
-            tweets.includes.media.find((media) => media.media_key === key)
-          ) || [],
-        referenced_tweets: getReferencedTweets(tweet),
-        author: getAuthorInfo(tweet.author_id),
-      };
-
-      return [tweetWithAuthor, ...allTweets];
-    }, []) || [] // If the Twitter API key isn't set, don't break the build
-  );
-};
-
-export const getTweetCount = async () => {
-  const response = await fetch(
-    `https://api.twitter.com/2/users/by/username/leeerob?user.fields=public_metrics`,
-    {
-      headers: {
-        Authorization: `Bearer ${process.env.TWITTER_API_TOKEN}`,
-      },
-    }
-  );
-
-  const { data } = await response.json();
-  return Number(data.public_metrics.tweet_count);
-};
