--- pages/api/auth/[...nextauth].ts
@@ -1,13 +1,0 @@
-import NextAuth, { NextAuthOptions } from 'next-auth';
-import GithubProvider from 'next-auth/providers/github';
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    GithubProvider({
-      clientId: process.env.OAUTH_CLIENT_KEY as string,
-      clientSecret: process.env.OAUTH_CLIENT_SECRET as string,
-    }),
-  ],
-};
-
-export default NextAuth(authOptions);
