--- pages/api/views/index.ts
@@ -1,19 +1,0 @@
-import { queryBuilder } from 'lib/planetscale';
-import type { NextApiRequest, NextApiResponse } from 'next';
-
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse
-) {
-  try {
-    const data = await queryBuilder
-      .selectFrom('views')
-      .select(['slug', 'count'])
-      .execute();
-
-    return res.status(200).json(data);
-  } catch (e) {
-    console.log(e);
-    return res.status(500).json({ message: e.message });
-  }
-}
