+++ types/projects.ts
@@ -1,13 +1,13 @@
 'use client';
 
-export type skillType = React.ReactElement;
+export type skillType = any;
 
 export type entryType = {
-	title: string;
-	titleIcon?: React.ReactElement;
-	workPeriod: string;
-	desc: string;
-	skills: skillType[];
-	id: string;
-	href: string;
+  title: string;
+  titleIcon?: React.ReactElement;
+  workPeriod: string;
+  desc: string;
+  skills: skillType[];
+  id: string;
+  href: string;
 };
